python_api/requirements.txt
gitpython
flask
flask-cors
pymysql
internetarchive
bs4
pypdf2
